+++ src/SatelliteCoverage/config.py
@@ -75,7 +75,7 @@ def compile_data(raw_paths):
 
         temp_df.rename(columns = {'sLat':'lat', 'sLon':'lon'}, inplace = True)
 
-        df = df.append(temp_df)
+        df = pd.concat([df,temp_df])
 
         # Updating counter
         i += 1
@@ -200,7 +200,7 @@ def filter_data(Date_options = None, IO = None, Metadata = None):
     summer_months = [6, 7, 8, 9, 10]
 
     date_path = IO['data_folder']
-    satellite = Metadata['ice_tracker']
+    satellite = Metadata['icetracker']
 
     if satellite == 'RCMS1':
         sat_list = ['rcm/','s1/']
@@ -263,279 +263,3 @@ def filter_data(Date_options = None, IO = None, Metadata = None):
 
     return raw_paths
 
-
-
-"""
-netCDF Analysis
-"""
-# Converts desired time range to seconds for netCDF analysis
-def date_to_seconds(path:str, start_year, start_month, start_day, end_year, end_month, end_day):
-    """
-    This function takes in a netCDF's reference time and start and end times of each triangle
-    and calculates the seconds elapsed between the former and latter. This is done to filter
-    the data loaded from the netCDF by time, as the start and end times in the netCDF are stored
-    as "seconds after the reference time".
-
-    INPUTS:
-    path -- String of path to the netCDF file the data will be read from {str}
-
-    start_year -- Start year of the user-specified period (Hereinafter the "period") {str}
-    start_month -- Start month of the period {str}
-    start_day -- Start day of the period {str}
-
-    end_year -- End year of the period {str}
-    end_month -- End month of the period {str}
-    end_day -- End day of the period {str}
-    """
-
-    # Open dataset
-    ds = Dataset(path, mode='r')
-
-    # Fetch reference time (start timestamp)
-    reftime = ds.getncattr('referenceTime')
-
-    # Closing dataset
-    ds.close()
-
-    # Converting reference time from string to datetime object
-    reftime = reftime[0:10]
-    reftime = datetime.strptime(reftime, '%Y-%m-%d')
-
-    # Converting start and end date strings to datetime objects
-    start_date = datetime.strptime(start_year + start_month + start_day, '%Y%m%d')
-    end_date = datetime.strptime(end_year + end_month + end_day, '%Y%m%d')
-
-    # Taking the difference between the start and end times of each triangle
-    # and converting it to seconds
-    start_ref_dt = (start_date - reftime).total_seconds()
-    end_ref_dt = (end_date - reftime).total_seconds()
-
-    return start_ref_dt, end_ref_dt
-
-# Converts date in seconds from netCDF's reference time to datetime object
-def seconds_to_date(path:str, date_sec_in):
-    # Open dataset
-    ds = Dataset(path, mode='r')
-
-    # Fetch reference time (start timestamp)
-    reftime = ds.getncattr('referenceTime')
-
-    # Closing dataset
-    ds.close()
-
-    # Converting reference time from string to datetime object
-    reftime = reftime[0:10]
-    reftime = datetime.strptime(reftime, '%Y-%m-%d')
-
-    date_out = reftime+timedelta(seconds=int(date_sec_in))
-
-    return date_out
-
-
-# Filters netCDF data by area
-def filter_area(centre_lat, centre_lon, radius, start_lats, start_lons):
-    """
-    This function filters data read from a netCDF (Stored as NumPy arrays) by area.
-    The user specifies a centre point in WGS 84 (Lat/Lon) and a radius in km, which are then
-    combined to create a mask, leaving only data *radius* kilometres from the centre point.
-    Only triangles with all three vertices within this circle will be loaded.
-
-    INPUTS:
-    centre_lat -- Latitude of the centre point {float}
-    centre_lon -- Longitude of the centre point {float}
-    radius -- Radius of the circle (mask) in kilometres {float}
-    start_lats -- Starting latitudes of the triangles in the format
-                  [[Latitudes (1)], [Latitudes (2)], [Latitudes (3)]] {NumPy array}
-    start_lons -- Starting latitudes of the triangles in the format
-                  [[Longitudes (1)], [Longitudes (2)], [Longitudes (3)]] {NumPy array}
-
-    OUTPUTS:
-    indices -- List of indices of triangles located within the area {list}
-    """
-
-    print('--- Filtering area ---')
-
-    # Turning inputs into floats if they aren't already
-    centre_lat, centre_lon, radius = float(centre_lat), float(centre_lon), float(radius)
-    centre_point = (centre_lat, centre_lon)
-
-    # Loading start lat/lon values
-    start_lat1, start_lat2, start_lat3 = start_lats[0], start_lats[1], start_lats[2]
-    start_lon1, start_lon2, start_lon3 = start_lons[0], start_lons[1], start_lons[2]
-
-    # List of 1s and 0s, 1 if the point is within the radius and 0 else
-    tf_list1, tf_list2, tf_list3 = ([] for i in range(3))
-
-    # Iterating over all lists
-    for coordinate in zip(start_lat1, start_lon1):
-        if hs.haversine(coordinate, centre_point) <= radius: # hs.haversine calculates the
-            tf_list1.append(1)                               # distance between two lats/lons
-        else:
-            tf_list1.append(0)
-
-    for coordinate in zip(start_lat2, start_lon2):
-        if hs.haversine(coordinate, centre_point) <= radius:
-            tf_list2.append(1)
-        else:
-            tf_list2.append(0)
-
-    for coordinate in zip(start_lat3, start_lon3):
-        if hs.haversine(coordinate, centre_point) <= radius:
-            tf_list3.append(1)
-        else:
-            tf_list3.append(0)
-
-    # Filtering indices where all three points (entire triangle) is within the radius
-    indices = [i for i in range(len(tf_list1)) if tf_list1[i] == tf_list2[i] == tf_list3[i] == 1]
-
-    return indices
-
-# Loads netCDF data
-def load_netcdf(path:str):
-    """
-    This function reads and loads data from a netCDF file in the same format as those output by the deformation
-    calculation script in src/SeaIceDeformation. The user is able to filter the data by time and area,
-    allowing for analytical tools to be applied to selected snippets of data.
-
-    INPUTS:
-    path -- String of path to the netCDF file the data will be read from {str}
-
-    OUTPUTS:
-    dictionary object -- Dictionary storing filtered data as NumPy arrays {dict}
-    """
-
-    print('--- Loading data ---')
-
-    # Reading config
-    config = read_config()
-    Date_options = config['Date_options']
-    options = config['options']
-
-    # Reading user options
-    start_year, start_month, start_day = Date_options['start_year'], Date_options['start_month'], Date_options['start_day']
-    end_year, end_month, end_day = Date_options['end_year'], Date_options['end_month'], Date_options['end_day']
-    area_filter, centre_lat, centre_lon, radius = options['area_filter'], options['centre_lat'], options['centre_lon'], options['radius']
-
-    # Load netCDF as Dataset from *path*
-    ds = Dataset(path, mode='r')
-
-    # Get start / end times from user
-    start_time_s, end_time_s = date_to_seconds(path, start_year, start_month, start_day, end_year, end_month, end_day)
-
-    # Extracting time variables
-    start_time = ds.variables['start_time'][:]
-    end_time = ds.variables['end_time'][:]
-
-    # Indices of data in desired time frame
-    time_indices = np.where( ((start_time < end_time_s) & (end_time > start_time_s)))[0]
-    start_time = start_time[time_indices]
-    end_time = end_time[time_indices]
-
-    # Extracting data (Filtered by time only)
-    start_lat1 = (ds.variables['start_lat1'][:])[time_indices]
-    start_lat2 = (ds.variables['start_lat2'][:])[time_indices]
-    start_lat3 = (ds.variables['start_lat3'][:])[time_indices]
-
-    start_lon1 = (ds.variables['start_lon1'][:])[time_indices]
-    start_lon2 = (ds.variables['start_lon2'][:])[time_indices]
-    start_lon3 = (ds.variables['start_lon3'][:])[time_indices]
-
-    end_lat1 = (ds.variables['end_lat1'][:])[time_indices]
-    end_lat2 = (ds.variables['end_lat2'][:])[time_indices]
-    end_lat3 = (ds.variables['end_lat3'][:])[time_indices]
-
-    end_lon1 = (ds.variables['end_lon1'][:])[time_indices]
-    end_lon2 = (ds.variables['end_lon2'][:])[time_indices]
-    end_lon3 = (ds.variables['end_lon3'][:])[time_indices]
-
-    div = (ds.variables['div'][:])[time_indices]
-    shr = (ds.variables['shr'][:])[time_indices]
-    vrt = (ds.variables['vrt'][:])[time_indices]
-
-    idx1 = (ds.variables['idx1'][:])[time_indices]
-    idx2 = (ds.variables['idx2'][:])[time_indices]
-    idx3 = (ds.variables['idx3'][:])[time_indices]
-    no   = (ds.variables['no'][:])[time_indices]
-
-    # DR: issue #11: This is not needed as idX is the same as id_start_latX
-    # id_start_lat1 = (ds.variables['id_start_lat1'][:])[time_indices]
-    # id_start_lat2 = (ds.variables['id_start_lat2'][:])[time_indices]
-    # id_start_lat3 = (ds.variables['id_start_lat3'][:])[time_indices]
-
-    dudx = (ds.variables['dudx'][:])[time_indices]
-    dudy = (ds.variables['dudy'][:])[time_indices]
-    dvdx = (ds.variables['dvdx'][:])[time_indices]
-    dvdy = (ds.variables['dvdy'][:])[time_indices]
-
-    min_date = seconds_to_date(path,np.min(start_time))
-    max_date = seconds_to_date(path,np.max(end_time))
-    print(f"Earliest/latest start/end dates of data included in deformation map: {min_date} to {max_date}")
-
-    # Compressing coordinates into arrays of arrays
-    # DR compression is not practical since it is not compressed in the NETcdf
-    # start_lats = np.array([start_lat1, start_lat2, start_lat3])
-    # start_lons = np.array([start_lon1, start_lon2, start_lon3])
-    # end_lats = np.array([end_lat1, end_lat2, end_lat3])
-    # end_lons = np.array([end_lon1, end_lon2, end_lon3])
-
-    reftime = ds.getncattr('referenceTime')
-    icetracker = ds.getncattr('iceTracker')
-    timestep = ds.getncattr('timestep')
-    tolerance = ds.getncattr('tolerance')
-    trackingerror = ds.getncattr('trackingError')
-
-    if area_filter == 'True':
-        print('--- Filtering by area ---')
-        # Filtering by area
-        area_indices = filter_area(centre_lat, centre_lon, radius, start_lats, start_lons)
-
-        # Applying filter
-        start_time = start_time[area_indices]
-        end_time = end_time[area_indices]
-
-        start_lat1 = start_lat1[area_indices]
-        start_lat2 = start_lat2[area_indices]
-        start_lat3 = start_lat3[area_indices]
-        start_lon1 = start_lon1[area_indices]
-        start_lon2 = start_lon2[area_indices]
-        start_lon3 = start_lon3[area_indices]
-
-        end_lat1 = end_lat1[area_indices]
-        end_lat2 = end_lat2[area_indices]
-        end_lat3 = end_lat3[area_indices]
-        end_lon1 = end_lon1[area_indices]
-        end_lon2 = end_lon2[area_indices]
-        end_lon3 = end_lon3[area_indices]
-
-        div = div[area_indices]
-        shr = shr[area_indices]
-        vrt = vrt[area_indices]
-
-        idx1 = idx1[area_indices]
-        idx2 = idx2[area_indices]
-        idx3 = idx3[area_indices]
-        no   = no[area_indices]
-
-        # DR: issue #11: This is not needed as idX is the same as id_start_latX
-        # id_start_lat1 = id_start_lat1[area_indices]
-        # id_start_lat2 = id_start_lat2[area_indices]
-        # id_start_lat3 = id_start_lat3[area_indices]
-
-        dudx = dudx[area_indices]
-        dudy = dudy[area_indices]
-        dvdx = dvdx[area_indices]
-        dvdy = dvdy[area_indices]
-
-    # Closing dataset
-    ds.close()
-
-    return {'start_lat1': start_lat1, 'start_lon1': start_lon1, 'end_lats': end_lat1, 'end_lons': end_lon,
-            'start_lat2': start_lat2, 'start_lon2': start_lon2, 'end_lats': end_lat2, 'end_lons': end_lon2,
-            'start_lat3': start_lat3, 'start_lon3': start_lon3, 'end_lats': end_lat3, 'end_lons': end_lon3,
-            'div': div, 'shr': shr, 'vrt': vrt,
-            'start_time': start_time, 'end_time': end_time, 'time_indices': time_indices,
-            'reftime': reftime, 'icetracker': icetracker, 'timestep': timestep,'tolerance': tolerance,
-            'trackingerror': trackingerror,
-            'idx1': idx1, 'idx2': idx2, 'idx3': idx3, 'no': no,
-            # 'start_id1': id_start_lat1, 'start_id2': id_start_lat2, 'start_id3': id_start_lat3,
-            'dudx': dudx, 'dudy': dudy, 'dvdx': dvdx, 'dvdy': dvdy}
