+++ src/SatelliteCoverage/options.ini
@@ -10,7 +10,7 @@
 #   netcdf_path:    Absolute path to netCDF file {str}
 #
 # [meta]
-#   ice_tracker:    Name of satellite(s) {str}
+#   icetracker:    Name of satellite(s) {str}
 #
 # [options]
 #   start_year:     Start year of analysis {int}
@@ -58,24 +58,25 @@
 
 
 [IO]
-data_folder   = /storage2/common/S1_RCM_tracked_pairs/
-output_folder = /storage2/amelie/ice-tracker-deformations/outputs/
-netcdf_path   = /storage2/common/S1_RCM_datasets/RCMS1SID_20171001_20220331_dt24_tol8_dx.nc
+data_folder = /storage/dringeisen/S1_RCM_tracked_pairs/
+output_folder = /storage/dringeisen/S1_RCM_tracked_pairs/outputs/
+netcdf_path   = /storage/dringeisen/S1_RCM_tracked_pairs/outputs/TESTING_dringeisen/05_output/RCMS1SID_20200301_20200302_dt72_tol5_dx.nc
+exp = TESTING_dringeisen
 
 [Metadata]
-ice_tracker = RCMS1
+icetracker = RCM
 
 [Date_options]
-start_year  = 2019
-start_month = 01
-start_day   = 12
+start_year  = 2020
+start_month = 03
+start_day   = 01
 
-end_year    = 2019
-end_month   = 01
-end_day     = 14
+end_year    = 2020
+end_month   = 03
+end_day     = 02
 
-timestep    = 24
-tolerance   = 8
+timestep    = 72
+tolerance   = 5
 
 [options]
 interval    = 24
