+++ src/SatelliteCoverage/utils.py
@@ -19,6 +19,16 @@ import haversine as hs
 import pandas as pd
 import pyproj
 
+import cartopy.crs as ccrs
+import cartopy.feature as cfeature
+from matplotlib.colors import Normalize
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+from datetime import datetime
+import numpy as np
+import pandas as pd
+import math
+
 def stb(s):
     if s in ['yes','Yes','true','True']:
          return True
@@ -45,11 +55,78 @@ def convert_to_grid(lon, lat):
     """
 
     # Input projection (lat/lon)
-    in_proj = pyproj.Proj(init='epsg:4326')
+    in_proj = pyproj.CRS('epsg:4326')
 
     # Output projection (EPSG 3413, Polar Stereographic)
-    out_proj = pyproj.Proj('+proj=stere +lat_0=90 +lat_ts=70 +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs ', preserve_units=True)
+    out_proj = pyproj.CRS('+proj=stere +lat_0=90 +lat_ts=70 +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs ', preserve_units=True)
 
     # Transform function
-    x, y = np.array(pyproj.transform(in_proj, out_proj, lon, lat))
+    y, x = np.array(pyproj.Transformer.from_crs(in_proj, out_proj).transform(lat, lon))
     return x, y
+
+"""
+netCDF Analysis
+"""
+# Converts desired time range to seconds for netCDF analysis
+def date_to_seconds(path:str, start_year, start_month, start_day, end_year, end_month, end_day):
+    """
+    This function takes in a netCDF's reference time and start and end times of each triangle
+    and calculates the seconds elapsed between the former and latter. This is done to filter
+    the data loaded from the netCDF by time, as the start and end times in the netCDF are stored
+    as "seconds after the reference time".
+
+    INPUTS:
+    path -- String of path to the netCDF file the data will be read from {str}
+
+    start_year -- Start year of the user-specified period (Hereinafter the "period") {str}
+    start_month -- Start month of the period {str}
+    start_day -- Start day of the period {str}
+
+    end_year -- End year of the period {str}
+    end_month -- End month of the period {str}
+    end_day -- End day of the period {str}
+    """
+
+    # Open dataset
+    ds = Dataset(path, mode='r')
+
+    # Fetch reference time (start timestamp)
+    reftime = ds.getncattr('referenceTime')
+
+    # Closing dataset
+    ds.close()
+
+    # Converting reference time from string to datetime object
+    reftime = reftime[0:10]
+    reftime = datetime.strptime(reftime, '%Y-%m-%d')
+
+    # Converting start and end date strings to datetime objects
+    start_date = datetime.strptime(start_year + start_month + start_day, '%Y%m%d')
+    end_date = datetime.strptime(end_year + end_month + end_day, '%Y%m%d')
+
+    # Taking the difference between the start and end times of each triangle
+    # and converting it to seconds
+    start_ref_dt = (start_date - reftime).total_seconds()
+    end_ref_dt = (end_date - reftime).total_seconds()
+
+    return start_ref_dt, end_ref_dt
+
+# Converts date in seconds from netCDF's reference time to datetime object
+def seconds_to_date(path:str, date_sec_in):
+    # Open dataset
+    ds = Dataset(path, mode='r')
+
+    # Fetch reference time (start timestamp)
+    reftime = ds.getncattr('referenceTime')
+
+    # Closing dataset
+    ds.close()
+
+    # Converting reference time from string to datetime object
+    reftime = reftime[0:10]
+    reftime = datetime.strptime(reftime, '%Y-%m-%d')
+
+    date_out = reftime+timedelta(seconds=int(date_sec_in))
+
+    return date_out
+
