+++ src/SeaIceDeformation/namelist.ini
@@ -25,7 +25,7 @@
 # ______________________________________________________________________________________________
 # Metadata
 #
-#   ice_tracker:    (S1/RCM) Sea-ice motion tracker
+#   icetracker:    (S1/RCM) Sea-ice motion tracker
 #
 #   tracking_error: Resolution or tracking error (meters) of te sea-ice motion tracker.
 #
@@ -73,7 +73,7 @@ bathymetry = /home/socn000/env_ubuntu-18.04-skylake-64/eccc-ppp4/datafiles/const
 cropped_grid = ./data/00_grid/cropped_grid.nc
 
 [Metadata]
-ice_tracker = RCMS1
+icetracker = RCMS1
 tracking_error = 200
 
 [Processing_options]
